requirements.txt
openai
requests
chardet
progressbar2